+++ crates/torin/src/lib.rs
@@ -1,6 +1,8 @@
 pub mod custom_measurer;
 pub mod dom_adapter;
 pub mod geometry;
+mod measure;
+mod measure_mode;
 pub mod node;
 pub mod scaled;
 pub mod torin;
