+++ crates/torin/src/torin.rs
@@ -6,12 +6,10 @@ use tracing::info;
 
 use crate::{
     custom_measurer::LayoutMeasurer,
-    direction::DirectionMode,
     dom_adapter::{DOMAdapter, NodeAreas, NodeKey},
     geometry::{Area, Size2D},
-    node::Node,
-    prelude::{Alignment, Gaps},
-    size::Size,
+    measure::measure_node,
+    prelude::Gaps,
 };
 
 /// Contains the best Root node candidate from where to start measuring
@@ -315,489 +313,4 @@ impl<Key: NodeKey> Torin<Key> {
     pub fn cache_node(&mut self, node_id: Key, areas: NodeAreas) {
         self.results.insert(node_id, areas);
     }
-}
-
-/// Measure this node and all it's children
-/// The caller of this function is responsible of caching the Node's layout results
-#[allow(clippy::too_many_arguments)]
-#[inline(always)]
-fn measure_node<Key: NodeKey>(
-    node_id: Key,
-    node: &Node,
-    layout: &mut Torin<Key>,
-    parent_area: &Area,
-    available_parent_area: &Area,
-    measurer: &mut Option<impl LayoutMeasurer<Key>>,
-    must_cache: bool,
-    dom_adapter: &mut impl DOMAdapter<Key>,
-) -> (bool, NodeAreas) {
-    let must_run = layout.dirty.contains(&node_id) || layout.results.get(&node_id).is_none();
-    if must_run {
-        let horizontal_padding = node.padding.horizontal();
-        let vertical_padding = node.padding.vertical();
-
-        let mut area = Rect::new(
-            available_parent_area.origin,
-            Size2D::new(horizontal_padding, vertical_padding),
-        );
-
-        area.origin.x += node.margin.left();
-        area.origin.y += node.margin.top();
-
-        area.size.width = node.width.min_max(
-            area.size.width,
-            parent_area.size.width,
-            node.margin.left(),
-            node.margin.horizontal(),
-            &node.minimum_width,
-            &node.maximum_width,
-        );
-        area.size.height = node.height.min_max(
-            area.size.height,
-            parent_area.size.height,
-            node.margin.top(),
-            node.margin.vertical(),
-            &node.minimum_height,
-            &node.maximum_height,
-        );
-
-        // Custom measure
-        let skip_inner = if let Some(measurer) = measurer {
-            let custom_measure =
-                measurer.measure(node_id, node, &area, parent_area, available_parent_area);
-            if let Some(new_area) = custom_measure {
-                if Size::Inner == node.width {
-                    area.size.width = node.width.min_max(
-                        new_area.width(),
-                        parent_area.size.width,
-                        node.margin.left(),
-                        node.margin.horizontal(),
-                        &node.minimum_width,
-                        &node.maximum_width,
-                    );
-                }
-                if Size::Inner == node.height {
-                    area.size.height = node.height.min_max(
-                        new_area.height(),
-                        parent_area.size.height,
-                        node.margin.top(),
-                        node.margin.vertical(),
-                        &node.minimum_height,
-                        &node.maximum_height,
-                    );
-                }
-            }
-            custom_measure.is_some()
-        } else {
-            false
-        };
-
-        let mut inner_sizes = Size2D::default();
-
-        // Node's inner area
-        let mut inner_area = {
-            let mut inner_area = area;
-            if Size::Inner == node.width {
-                inner_area.size.width = available_parent_area.width()
-            }
-            if Size::Inner == node.height {
-                inner_area.size.height = available_parent_area.height()
-            }
-            inner_area
-        };
-
-        // Apply padding
-        inner_area.origin.x += node.padding.left();
-        inner_area.origin.y += node.padding.top();
-        inner_area.size.width -= horizontal_padding;
-        inner_area.size.height -= vertical_padding;
-
-        // Node's available inner area
-        let mut available_area = inner_area;
-
-        // Apply scroll
-        available_area.origin.x += node.offset_x.get();
-        available_area.origin.y += node.offset_y.get();
-
-        let mut measurement_mode = MeasureMode::ParentIsNotCached {
-            area: &mut area,
-            inner_area: &mut inner_area,
-            vertical_padding,
-            horizontal_padding,
-        };
-
-        if !skip_inner {
-            measure_inner_nodes(
-                &node_id,
-                node,
-                layout,
-                &mut available_area,
-                &mut inner_sizes,
-                measurer,
-                must_cache,
-                &mut measurement_mode,
-                dom_adapter,
-            );
-        }
-
-        (
-            must_cache,
-            NodeAreas {
-                area,
-                margin: node.margin,
-                inner_area,
-                inner_sizes,
-            },
-        )
-    } else {
-        let areas = layout.get(node_id).unwrap().clone();
-
-        let mut inner_sizes = areas.inner_sizes;
-        let mut available_area = areas.inner_area;
-
-        // TODO(marc2332): Should I also cache these?
-        available_area.origin.x += node.offset_x.get();
-        available_area.origin.y += node.offset_y.get();
-
-        let mut measurement_mode = MeasureMode::ParentIsCached {
-            inner_area: &areas.inner_area,
-        };
-
-        measure_inner_nodes(
-            &node_id,
-            node,
-            layout,
-            &mut available_area,
-            &mut inner_sizes,
-            measurer,
-            must_cache,
-            &mut measurement_mode,
-            dom_adapter,
-        );
-
-        (false, areas)
-    }
-}
-
-/// Measurement data for the inner Nodes of a Node
-#[derive(Debug)]
-enum MeasureMode<'a> {
-    ParentIsCached {
-        inner_area: &'a Area,
-    },
-    ParentIsNotCached {
-        area: &'a mut Area,
-        inner_area: &'a mut Area,
-        vertical_padding: f32,
-        horizontal_padding: f32,
-    },
-}
-
-impl<'a> MeasureMode<'a> {
-    /// Get a reference to the inner area
-    pub fn inner_area(&'a self) -> &'a Area {
-        match self {
-            Self::ParentIsCached { inner_area } => inner_area,
-            Self::ParentIsNotCached { inner_area, .. } => inner_area,
-        }
-    }
-}
-
-#[derive(Debug)]
-enum OwnedMeasureMode {
-    ParentIsCached {
-        inner_area: Area,
-    },
-    ParentIsNotCached {
-        area: Area,
-        inner_area: Area,
-        vertical_padding: f32,
-        horizontal_padding: f32,
-    },
-}
-
-impl OwnedMeasureMode {
-    pub fn ref_mut(&mut self) -> MeasureMode<'_> {
-        match self {
-            Self::ParentIsCached { inner_area } => MeasureMode::ParentIsCached { inner_area },
-            Self::ParentIsNotCached {
-                area,
-                inner_area,
-                vertical_padding,
-                horizontal_padding,
-            } => MeasureMode::ParentIsNotCached {
-                area,
-                inner_area,
-                vertical_padding: *vertical_padding,
-                horizontal_padding: *horizontal_padding,
-            },
-        }
-    }
-}
-
-impl MeasureMode<'_> {
-    pub fn to_owned(&self) -> OwnedMeasureMode {
-        match self {
-            MeasureMode::ParentIsCached { inner_area } => OwnedMeasureMode::ParentIsCached {
-                inner_area: *inner_area.to_owned(),
-            },
-            MeasureMode::ParentIsNotCached {
-                area,
-                inner_area,
-                vertical_padding,
-                horizontal_padding,
-            } => OwnedMeasureMode::ParentIsNotCached {
-                area: **area.clone(),
-                inner_area: **inner_area.clone(),
-                vertical_padding: *vertical_padding,
-                horizontal_padding: *horizontal_padding,
-            },
-        }
-    }
-}
-
-/// Measure the inner Nodes of a Node
-#[allow(clippy::too_many_arguments)]
-#[inline(always)]
-fn measure_inner_nodes<Key: NodeKey>(
-    node_id: &Key,
-    node: &Node,
-    layout: &mut Torin<Key>,
-    available_area: &mut Area,
-    inner_sizes: &mut Size2D,
-    measurer: &mut Option<impl LayoutMeasurer<Key>>,
-    must_cache: bool,
-    mode: &mut MeasureMode,
-    dom_adapter: &mut impl DOMAdapter<Key>,
-) {
-    let mut measure_children = |mode: &mut MeasureMode,
-                                available_area: &mut Area,
-                                inner_sizes: &mut Size2D,
-                                measure_cross_axis: bool,
-                                must_cache: bool| {
-        let children = dom_adapter.children_of(node_id);
-
-        for child_id in children {
-            let inner_area = *mode.inner_area();
-
-            let child_data = dom_adapter.get_node(&child_id).unwrap();
-
-            let mut adapted_available_area = *available_area;
-
-            // If required, make a previous measure to adapt the position
-            if measure_cross_axis {
-                let (_, child_areas) = measure_node(
-                    child_id,
-                    &child_data,
-                    layout,
-                    &inner_area,
-                    available_area,
-                    measurer,
-                    false,
-                    dom_adapter,
-                );
-
-                match node.direction {
-                    DirectionMode::Horizontal => match node.cross_alignment {
-                        Alignment::Center => {
-                            let new_origin_y =
-                                (available_area.height() / 2.0) - (child_areas.area.height() / 2.0);
-
-                            adapted_available_area.origin.y = available_area.min_y() + new_origin_y;
-                        }
-                        Alignment::End => {
-                            adapted_available_area.origin.y =
-                                available_area.max_y() - child_areas.area.height();
-                        }
-                        _ => {}
-                    },
-                    DirectionMode::Vertical => match node.cross_alignment {
-                        Alignment::Center => {
-                            let new_origin_x =
-                                (available_area.width() / 2.0) - (child_areas.area.width() / 2.0);
-
-                            adapted_available_area.origin.x = available_area.min_x() + new_origin_x;
-                        }
-                        Alignment::End => {
-                            adapted_available_area.origin.x =
-                                available_area.max_x() - child_areas.area.width();
-                        }
-                        _ => {}
-                    },
-                }
-            }
-
-            let (child_revalidated, child_areas) = measure_node(
-                child_id,
-                &child_data,
-                layout,
-                &inner_area,
-                &adapted_available_area,
-                measurer,
-                must_cache,
-                dom_adapter,
-            );
-
-            match node.direction {
-                DirectionMode::Horizontal => {
-                    // Move the available area
-                    available_area.origin.x = child_areas.area.max_x();
-                    available_area.size.width -= child_areas.area.size.width;
-
-                    if let MeasureMode::ParentIsNotCached {
-                        area,
-                        vertical_padding,
-                        inner_area,
-                        ..
-                    } = mode
-                    {
-                        inner_sizes.height = child_areas.area.height().max(inner_sizes.height);
-                        inner_sizes.width += child_areas.area.width();
-
-                        // Keep the biggest height
-                        if node.height == Size::Inner {
-                            area.size.height = area
-                                .size
-                                .height
-                                .max(child_areas.area.size.height + *vertical_padding);
-                            // Keep the inner area in sync
-                            inner_area.size.height = area.size.height - *vertical_padding;
-                        }
-
-                        // Accumulate width
-                        if node.width == Size::Inner {
-                            area.size.width += child_areas.area.size.width;
-                        }
-                    }
-                }
-                DirectionMode::Vertical => {
-                    // Move the available area
-                    available_area.origin.y = child_areas.area.max_y();
-                    available_area.size.height -= child_areas.area.size.height;
-
-                    if let MeasureMode::ParentIsNotCached {
-                        area,
-                        horizontal_padding,
-                        inner_area,
-                        ..
-                    } = mode
-                    {
-                        inner_sizes.width = child_areas.area.width().max(inner_sizes.width);
-                        inner_sizes.height += child_areas.area.height();
-
-                        // Keep the biggest width
-                        if node.width == Size::Inner {
-                            area.size.width = area
-                                .size
-                                .width
-                                .max(child_areas.area.size.width + *horizontal_padding);
-                            // Keep the inner area in sync
-                            inner_area.size.width = area.size.width - *horizontal_padding;
-                        }
-
-                        // Accumulate height
-                        if node.height == Size::Inner {
-                            area.size.height += child_areas.area.size.height;
-                        }
-                    }
-                }
-            }
-
-            if child_revalidated && must_cache {
-                layout.cache_node(child_id, child_areas);
-            }
-        }
-    };
-
-    // If any of the alignments isn't "start", it needs to run a measurement to get the aligned position, before caching the value
-    let has_special_alignments =
-        node.main_alignment.is_not_start() || node.cross_alignment.is_not_start();
-
-    {
-        let mut alignment_mode = mode.to_owned();
-        let mut alignment_mode = alignment_mode.ref_mut();
-        let mut inner_sizes = *inner_sizes;
-
-        if has_special_alignments {
-            measure_children(
-                &mut alignment_mode,
-                &mut available_area.clone(),
-                &mut inner_sizes,
-                true,
-                false,
-            );
-
-            match node.direction {
-                DirectionMode::Horizontal => match node.main_alignment {
-                    Alignment::Center => {
-                        let inner_area = alignment_mode.inner_area();
-                        let new_origin_x = (inner_area.width() / 2.0) - (inner_sizes.width / 2.0);
-
-                        available_area.origin.x = inner_area.min_x() + new_origin_x;
-                    }
-                    Alignment::End => {
-                        let inner_area = alignment_mode.inner_area();
-                        available_area.origin.x = inner_area.max_x() - inner_sizes.width;
-                    }
-                    _ => {}
-                },
-                DirectionMode::Vertical => {
-                    // If the height is auto, we set the inner area to the parent area
-                    // TODO: CLEAN THIS UP LOL
-                    if Size::Inner == node.height && node.main_alignment.is_not_start() {
-                        if let MeasureMode::ParentIsNotCached {
-                            area,
-                            inner_area,
-                            vertical_padding,
-                            ..
-                        } = &mut alignment_mode
-                        {
-                            inner_area.origin.y = area.origin.y + node.padding.top();
-                            inner_area.size.height = area.size.height - *vertical_padding;
-                            available_area.size.height = inner_area.size.height;
-                        }
-                    }
-
-                    // If the width is auto, we set the inner area to the parent area
-                    // TODO: CLEAN THIS UP LOL
-                    if Size::Inner == node.width && node.cross_alignment.is_not_start() {
-                        if let MeasureMode::ParentIsNotCached {
-                            area,
-                            inner_area,
-                            horizontal_padding,
-                            ..
-                        } = &mut alignment_mode
-                        {
-                            inner_area.origin.x = area.origin.x + node.padding.left();
-                            inner_area.size.width = area.size.width - *horizontal_padding;
-                            available_area.size.width = inner_area.size.width;
-                        }
-                    }
-
-                    match node.main_alignment {
-                        Alignment::Center => {
-                            let inner_area = alignment_mode.inner_area();
-                            let new_origin_y =
-                                (inner_area.height() / 2.0) - (inner_sizes.height / 2.0);
-
-                            available_area.origin.y = inner_area.min_y() + new_origin_y;
-                        }
-                        Alignment::End => {
-                            let inner_area = alignment_mode.inner_area();
-                            available_area.origin.y = inner_area.max_y() - inner_sizes.height;
-                        }
-                        _ => {}
-                    }
-                }
-            }
-        }
-    }
-
-    measure_children(
-        mode,
-        available_area,
-        inner_sizes,
-        has_special_alignments,
-        must_cache,
-    );
 }
